Add option to extend effective pressure for the velocity solver
This is needed because the friction laws that use effective pressure
need N on the first floating cell to allow for more accurate evaluation
of friction at quadrature points.
Also, this commit moves the limiting of the effective pressure from the
velocity_external routine to the calc_beta routine.
Finally, all these manipulations to N are done on the field
effectivePressureLimited to avoid messing up effectivePressure which is
used by the subglacial hydrology model (if on).